testing/20140819.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20140819</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="w3c/testing.20140819.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2014-08-19 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Tue Aug 19 00:00:01 2014</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #testing</li>
 <li id="l-3"><a href="#l-3">#</a> [00:18] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-4"><a href="#l-4">#</a> [00:20] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-5"><a href="#l-5">#</a> [00:28] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-6"><a href="#l-6">#</a> [00:48] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-7"><a href="#l-7">#</a> [00:50] * Joins: mdas_ (~mdas@public.cloak)</li>
 <li id="l-8"><a href="#l-8">#</a> [00:50] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-9"><a href="#l-9">#</a> [00:57] * Quits: mdas_ (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-10"><a href="#l-10">#</a> [01:51] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-11"><a href="#l-11">#</a> [01:58] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-12"><a href="#l-12">#</a> [02:01] * Quits: ptressel (~chatzilla@public.cloak) (&quot;bbl&quot;)</li>
 <li id="l-13"><a href="#l-13">#</a> [02:16] * Quits: Automate_ (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-14" _a="heycam|away"><a href="#l-14">#</a> [02:30] * heycam|away is now known as heycam</li>
 <li id="l-15"><a href="#l-15">#</a> [02:52] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-16"><a href="#l-16">#</a> [02:59] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-17"><a href="#l-17">#</a> [03:52] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-18"><a href="#l-18">#</a> [03:59] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-19"><a href="#l-19">#</a> [04:12] * Joins: mdyck (~mdyck@public.cloak)</li>
 <li id="l-20" _a="heycam"><a href="#l-20">#</a> [04:33] * heycam is now known as heycam|away</li>
 <li id="l-21"><a href="#l-21">#</a> [04:53] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-22"><a href="#l-22">#</a> [04:55] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-23"><a href="#l-23">#</a> [04:55] -gitbot:#testing- [web-platform-tests] sideshowbarker closed pull request #463: Import latest tests from html5lib-tests (master...new-ruby) <a href="https://github.com/w3c/web-platform-tests/pull/463">https://github.com/w3c/web-platform-tests/pull/463</a></li>
 <li id="l-24"><a href="#l-24">#</a> [04:55] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-25"><a href="#l-25">#</a> [04:55] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-26"><a href="#l-26">#</a> [04:55] -gitbot:#testing- [web-platform-tests] sideshowbarker deleted new-ruby at f3eea8d: <a href="https://github.com/w3c/web-platform-tests/commit/f3eea8d">https://github.com/w3c/web-platform-tests/commit/f3eea8d</a></li>
 <li id="l-27"><a href="#l-27">#</a> [04:55] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-28"><a href="#l-28">#</a> [05:00] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-29"><a href="#l-29">#</a> [05:54] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-30"><a href="#l-30">#</a> [06:01] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-31"><a href="#l-31">#</a> [06:11] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-32"><a href="#l-32">#</a> [06:18] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-33"><a href="#l-33">#</a> [07:01] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-34"><a href="#l-34">#</a> [07:08] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-35"><a href="#l-35">#</a> [07:29] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-36"><a href="#l-36">#</a> [08:01] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-37"><a href="#l-37">#</a> [08:08] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-38"><a href="#l-38">#</a> [08:43] * Joins: zcorpa___ (~zcorpan@public.cloak)</li>
 <li id="l-39"><a href="#l-39">#</a> [08:53] * Joins: dom (dom@public.cloak)</li>
 <li id="l-40"><a href="#l-40">#</a> [08:53] * Joins: anssik (~uid10742@public.cloak)</li>
 <li id="l-41"><a href="#l-41">#</a> [09:02] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-42"><a href="#l-42">#</a> [09:09] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-43"><a href="#l-43">#</a> [09:16] * Quits: mdyck (~mdyck@public.cloak) (Client closed connection)</li>
 <li id="l-44"><a href="#l-44">#</a> [09:33] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-45"><a href="#l-45">#</a> [09:41] * Quits: ptressel (~chatzilla@public.cloak) (&quot;zzz&quot;)</li>
 <li id="l-46"><a href="#l-46">#</a> [09:54] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-47"><a href="#l-47">#</a> [09:57] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-48"><a href="#l-48">#</a> [10:03] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-49"><a href="#l-49">#</a> [10:10] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-50"><a href="#l-50">#</a> [10:34] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-51" _a="jgraham" _r="Woah"><a href="#l-51">#</a> [10:58] &lt;jgraham&gt; Woah, CSS reftests have a complex naming convention for reference files but *also* require you to add a &lt;link&gt; element?</li>
 <li id="l-52" _a="jgraham"><a href="#l-52">#</a> [11:00] &lt;jgraham&gt; I assumed it was one or the other until I read more documentation</li>
 <li id="l-53"><a href="#l-53">#</a> [11:02] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-54"><a href="#l-54">#</a> [11:04] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-55"><a href="#l-55">#</a> [11:05] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-56" _a="zcorpa___"><a href="#l-56">#</a> [11:08] &lt;zcorpa___&gt; moar metadata</li>
 <li id="l-57"><a href="#l-57">#</a> [11:11] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-58"><a href="#l-58">#</a> [11:14] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-59" _a="jgraham"><a href="#l-59">#</a> [11:23] &lt;jgraham&gt; Is there any software other than shepherd that supports this multiple linked references idea?</li>
 <li id="l-60" _a="Ms2ger"><a href="#l-60">#</a> [11:25] &lt;Ms2ger&gt; Hmm?</li>
 <li id="l-61" _a="Ms2ger"><a href="#l-61">#</a> [11:25] &lt;Ms2ger&gt; What idea is that?</li>
 <li id="l-62" _a="jgraham"><a href="#l-62">#</a> [11:25] &lt;jgraham&gt; You have a ref file with a &lt;link&gt; to another ref file and so on and all references have to match for the test to pass</li>
 <li id="l-63" _a="Ms2ger"><a href="#l-63">#</a> [11:26] &lt;Ms2ger&gt; Mm</li>
 <li id="l-64" _a="Ms2ger"><a href="#l-64">#</a> [11:26] &lt;Ms2ger&gt; That sounds kinda neat</li>
 <li id="l-65" _a="jgraham" _r="Yeah"><a href="#l-65">#</a> [11:27] &lt;jgraham&gt; Yeah, although it fundamentally changes the data model</li>
 <li id="l-66" _a="jgraham"><a href="#l-66">#</a> [11:27] &lt;jgraham&gt; And it isn't *really* necessary; it's just an optimisation to collapse N tests down to 1 test</li>
 <li id="l-67" _a="Ms2ger"><a href="#l-67">#</a> [11:27] &lt;Ms2ger&gt; Yep</li>
 <li id="l-68" _a="jgraham"><a href="#l-68">#</a> [11:28] &lt;jgraham&gt; So apparently I didn't read any of the testthewebforward docs until I started rewriting them</li>
 <li id="l-69" _a="jgraham"><a href="#l-69">#</a> [11:28] &lt;jgraham&gt; As a side effect, I now have lots of unrelated ideas for fun things to implement that would be less boring than writing documentation</li>
 <li id="l-70" _a="Ms2ger"><a href="#l-70">#</a> [11:29] &lt;Ms2ger&gt; File issues, continue with the docs ;)</li>
 <li id="l-71" _a="jgraham"><a href="#l-71">#</a> [11:29] &lt;jgraham&gt; Dammit :p</li>
 <li id="l-72" _a="jgraham"><a href="#l-72">#</a> [11:30] &lt;jgraham&gt; But yeah I want the docs to be in a not-terrible state when I am convincing people at Mozilla that web-platform-tests are a good idea</li>
 <li id="l-73" _a="jgraham"><a href="#l-73">#</a> [11:31] &lt;jgraham&gt; Although I'm not sure I can fix it that fast :(</li>
 <li id="l-74" _a="jgraham"><a href="#l-74">#</a> [11:44] &lt;jgraham&gt; Sigh</li>
 <li id="l-75" _a="jgraham"><a href="#l-75">#</a> [11:44] &lt;jgraham&gt; There seems to be a fundamental semantic mismatch between the CSS tests and w-p-t on the subject of reftests</li>
 <li id="l-76" _a="zcorpa___"><a href="#l-76">#</a> [11:56] &lt;zcorpa___&gt; what is the mismatch?</li>
 <li id="l-77" _a="jgraham" _r="zcorpa___"><a href="#l-77">#</a> [12:01] &lt;jgraham&gt; zcorpa___: See my email to public-test-infra/public-css-testsuite</li>
 <li id="l-78"><a href="#l-78">#</a> [12:06] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-79"><a href="#l-79">#</a> [12:12] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-80"><a href="#l-80">#</a> [12:23] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-81" _a="abarsto"><a href="#l-81">#</a> [12:23] * abarsto is now known as ArtB</li>
 <li id="l-82" _a="MikeSmith" _r="zcorpa___"><a href="#l-82">#</a> [12:40] &lt;MikeSmith&gt; zcorpa___: blink bug for ruby changes, with a patch waiting for review <a href="https://code.google.com/p/chromium/issues/detail?id=404892">https://code.google.com/p/chromium/issues/detail?id=404892</a></li>
 <li id="l-83" _a="MikeSmith"><a href="#l-83">#</a> [12:41] &lt;MikeSmith&gt; ah wait I guess it's been reviewed already and committed <a href="https://codereview.chromium.org/482373002">https://codereview.chromium.org/482373002</a></li>
 <li id="l-84" _a="MikeSmith"><a href="#l-84">#</a> [12:44] &lt;MikeSmith&gt; <a href="http://src.chromium.org/viewvc/blink?view=revision&amp;revision=180523">http://src.chromium.org/viewvc/blink?view=revision&amp;revision=180523</a></li>
 <li id="l-85" _a="zcorpa___" _r="MikeSmith"><a href="#l-85">#</a> [12:46] &lt;zcorpa___&gt; MikeSmith: thx</li>
 <li id="l-86"><a href="#l-86">#</a> [13:05] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-87"><a href="#l-87">#</a> [13:12] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-88"><a href="#l-88">#</a> [13:27] * Joins: scott_gonzalez (~scott_gonzalez@public.cloak)</li>
 <li id="l-89"><a href="#l-89">#</a> [13:51] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-90"><a href="#l-90">#</a> [14:06] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-91"><a href="#l-91">#</a> [14:13] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-92" _a="zcorpa___" _r="MikeSmith"><a href="#l-92">#</a> [14:39] &lt;zcorpa___&gt; MikeSmith: i meant how does the html5lib test runner run tests with #script-off ?</li>
 <li id="l-93" _a="zcorpa___"><a href="#l-93">#</a> [14:39] * zcorpa___ is now known as zcorpan</li>
 <li id="l-94" _a="jgraham"><a href="#l-94">#</a> [14:39] &lt;jgraham&gt; I'm not sure it does</li>
 <li id="l-95" _a="jgraham"><a href="#l-95">#</a> [14:39] &lt;jgraham&gt; It's safe to use ascii irrespective of browser locale, right?</li>
 <li id="l-96" _a="zcorpan"><a href="#l-96">#</a> [14:40] &lt;zcorpan&gt; ask hsivonen</li>
 <li id="l-97" _a="zcorpan"><a href="#l-97">#</a> [14:40] &lt;zcorpan&gt; but i think yes</li>
 <li id="l-98" _a="MikeSmith" _r="zcorpan"><a href="#l-98">#</a> [14:46] &lt;MikeSmith&gt; zcorpan: how actually do I run the html5lib test runner run tests with #script-off ?</li>
 <li id="l-99" _a="zcorpan" _r="MikeSmith"><a href="#l-99">#</a> [14:47] &lt;zcorpan&gt; MikeSmith: i think it's not really possible</li>
 <li id="l-100" _a="MikeSmith"><a href="#l-100">#</a> [14:48] &lt;MikeSmith&gt; ah ok</li>
 <li id="l-101"><a href="#l-101">#</a> [14:49] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-102"><a href="#l-102">#</a> [14:49] -gitbot:#testing- [web-platform-tests] cptvitamin closed pull request #1197: fixed the test to work as expected (master...cptvitamin/drawFocusAAPItest) <a href="https://github.com/w3c/web-platform-tests/pull/1197">https://github.com/w3c/web-platform-tests/pull/1197</a></li>
 <li id="l-103"><a href="#l-103">#</a> [14:49] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-104" _a="jgraham"><a href="#l-104">#</a> [14:54] * jgraham wonders how much of <a href="http://testthewebforward.org/docs/test-format-guidelines.html#file-name-format">http://testthewebforward.org/docs/test-format-guidelines.html#file-name-format</a> is actually relevant</li>
 <li id="l-105"><a href="#l-105">#</a> [14:54] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-106" _a="jgraham"><a href="#l-106">#</a> [14:55] &lt;jgraham&gt; I mean I guess the CSS people care, but it clearly isn't enforced for wpt</li>
 <li id="l-107" _a="jgraham"><a href="#l-107">#</a> [14:55] &lt;jgraham&gt; (also that page is basically impossible to find)</li>
 <li id="l-108" _a="jgraham"><a href="#l-108">#</a> [14:55] &lt;jgraham&gt; I feel like I keep writing &quot;Note: CSS has an extra gotcha here&quot;</li>
 <li id="l-109" _a="MikeSmith" _r="zcorpan"><a href="#l-109">#</a> [15:04] &lt;MikeSmith&gt; zcorpan: so then is there still something more I need to do at this point for the &quot;Import latest tests from html5lib-tests&quot; PR?</li>
 <li id="l-110"><a href="#l-110">#</a> [15:05] * Quits: MK (kawamori@public.cloak) (&quot;Leaving...&quot;)</li>
 <li id="l-111"><a href="#l-111">#</a> [15:07] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-112"><a href="#l-112">#</a> [15:07] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-113"><a href="#l-113">#</a> [15:14] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-114"><a href="#l-114">#</a> [15:20] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-115"><a href="#l-115">#</a> [15:30] * Joins: MK_ (kawamori@public.cloak)</li>
 <li id="l-116"><a href="#l-116">#</a> [15:30] * Quits: MK_ (kawamori@public.cloak) (&quot;Leaving...&quot;)</li>
 <li id="l-117" _a="zcorpan" _r="MikeSmith"><a href="#l-117">#</a> [15:38] &lt;zcorpan&gt; MikeSmith: i think it introduces at least one #script-off test that will always fail - <a href="https://github.com/html5lib/html5lib-tests/blob/master/tree-construction/tests16.dat#L1042">https://github.com/html5lib/html5lib-tests/blob/master/tree-construction/tests16.dat#L1042</a></li>
 <li id="l-118" _a="zcorpan" _r="MikeSmith"><a href="#l-118">#</a> [15:39] &lt;zcorpan&gt; MikeSmith: so #script-off tests need to be removed for wpt i think</li>
 <li id="l-119" _a="zcorpan" _r="MikeSmith"><a href="#l-119">#</a> [15:49] &lt;zcorpan&gt; MikeSmith: compare <a href="http://w3c-test.org/submissions/1195/html/syntax/parsing/html5lib_tests16.html">http://w3c-test.org/submissions/1195/html/syntax/parsing/html5lib_tests16.html</a> and <a href="http://w3c-test.org/html/syntax/parsing/html5lib_tests16.html">http://w3c-test.org/html/syntax/parsing/html5lib_tests16.html</a> in firefox</li>
 <li id="l-120"><a href="#l-120">#</a> [16:34] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-121"><a href="#l-121">#</a> [16:34] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/file_idl_constant_name (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/efc9f069768f">https://github.com/w3c/web-platform-tests/commit/efc9f069768f</a></li>
 <li id="l-122"><a href="#l-122">#</a> [16:34] -gitbot:#testing- web-platform-tests/jgraham/file_idl_constant_name efc9f06 James Graham: Avoid bug in debug Firefox builds where FileAPI tests get a new name each run.</li>
 <li id="l-123"><a href="#l-123">#</a> [16:34] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-124"><a href="#l-124">#</a> [16:35] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-125"><a href="#l-125">#</a> [16:35] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1198: Avoid bug in debug Firefox builds where FileAPI tests get a new name each run. (master...jgraham/file_idl_constant_name) <a href="https://github.com/w3c/web-platform-tests/pull/1198">https://github.com/w3c/web-platform-tests/pull/1198</a></li>
 <li id="l-126"><a href="#l-126">#</a> [16:35] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-127" _a="ato"><a href="#l-127">#</a> [16:37] &lt;ato&gt; How do I assert that something is not undefined or none in testharness.js?</li>
 <li id="l-128" _a="ato"><a href="#l-128">#</a> [16:39] &lt;ato&gt; s/none/null/ I mean of course</li>
 <li id="l-129"><a href="#l-129">#</a> [16:39] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-130"><a href="#l-130">#</a> [16:39] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/010de9208b661bdd01891a6d3a2cffb4ef35b70b">https://github.com/w3c/web-platform-tests/commit/010de9208b661bdd01891a6d3a2cffb4ef35b70b</a></li>
 <li id="l-131"><a href="#l-131">#</a> [16:39] -gitbot:#testing- web-platform-tests/master 010de92 Ms2ger: Merge pull request #1198 from w3c/jgraham/file_idl_constant_name...</li>
 <li id="l-132"><a href="#l-132">#</a> [16:39] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-133"><a href="#l-133">#</a> [16:39] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-134"><a href="#l-134">#</a> [16:39] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #1198: Avoid bug in debug Firefox builds where FileAPI tests get a new name each run. (master...jgraham/file_idl_constant_name) <a href="https://github.com/w3c/web-platform-tests/pull/1198">https://github.com/w3c/web-platform-tests/pull/1198</a></li>
 <li id="l-135"><a href="#l-135">#</a> [16:39] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-136" _a="Ms2ger" _r="ato"><a href="#l-136">#</a> [16:40] &lt;Ms2ger&gt; ato, in general, assert something more specific</li>
 <li id="l-137"><a href="#l-137">#</a> [16:40] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-138"><a href="#l-138">#</a> [16:40] -gitbot:#testing- [web-platform-tests] Ms2ger deleted jgraham/file_idl_constant_name at efc9f06: <a href="https://github.com/w3c/web-platform-tests/commit/efc9f06">https://github.com/w3c/web-platform-tests/commit/efc9f06</a></li>
 <li id="l-139"><a href="#l-139">#</a> [16:40] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-140" _a="ato" _r="Ms2ger"><a href="#l-140">#</a> [16:41] &lt;ato&gt; Ms2ger: So assert_equals(typeof obj !== undefined) is the closest I get?</li>
 <li id="l-141" _a="ato" _r="Sorry"><a href="#l-141">#</a> [16:41] &lt;ato&gt; Sorry, I mean assert_true.</li>
 <li id="l-142" _a="Ms2ger"><a href="#l-142">#</a> [16:41] &lt;Ms2ger&gt; You can do that, yes</li>
 <li id="l-143" _a="Ms2ger"><a href="#l-143">#</a> [16:42] &lt;Ms2ger&gt; It's better to test what it is than what it isn't, though</li>
 <li id="l-144" _a="ato" _r="Alright"><a href="#l-144">#</a> [16:42] &lt;ato&gt; Alright, thanks.</li>
 <li id="l-145" _a="jgraham" _r="MikeSmith"><a href="#l-145">#</a> [16:48] &lt;jgraham&gt; MikeSmith, zcorpan: Did you decide what's happening with that html5lib update? I would like to bitrot you by removing some harmful legacy cruft to do with timeouts from the harness</li>
 <li id="l-146"><a href="#l-146">#</a> [16:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-147"><a href="#l-147">#</a> [16:59] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/html5lib_default_timeout (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/b1e419678d25">https://github.com/w3c/web-platform-tests/commit/b1e419678d25</a></li>
 <li id="l-148"><a href="#l-148">#</a> [16:59] -gitbot:#testing- web-platform-tests/jgraham/html5lib_default_timeout b1e4196 James Graham: Remove the legacy timeouts from html5lib tests.</li>
 <li id="l-149"><a href="#l-149">#</a> [16:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-150"><a href="#l-150">#</a> [16:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-151"><a href="#l-151">#</a> [16:59] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1199: Remove the legacy timeouts from html5lib tests. (master...jgraham/html5lib_default_timeout) <a href="https://github.com/w3c/web-platform-tests/pull/1199">https://github.com/w3c/web-platform-tests/pull/1199</a></li>
 <li id="l-152"><a href="#l-152">#</a> [16:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-153" _a="jgraham"><a href="#l-153">#</a> [17:00] &lt;jgraham&gt; r? <a href="https://critic.hoppipolla.co.uk/r/2398">https://critic.hoppipolla.co.uk/r/2398</a></li>
 <li id="l-154" _a="jgraham"><a href="#l-154">#</a> [17:00] &lt;jgraham&gt; Mostly generated changes, only a few actual changes</li>
 <li id="l-155" _a="zcorpan" _r="jgraham"><a href="#l-155">#</a> [17:02] &lt;zcorpan&gt; jgraham: <a href="https://critic.hoppipolla.co.uk/showcomment?chain=7302">https://critic.hoppipolla.co.uk/showcomment?chain=7302</a></li>
 <li id="l-156" _a="zcorpan" _r="jgraham"><a href="#l-156">#</a> [17:04] &lt;zcorpan&gt; jgraham: you don't need long timeout meta for fragment tests?</li>
 <li id="l-157" _a="jgraham" _r="zcorpan"><a href="#l-157">#</a> [17:07] &lt;jgraham&gt; zcorpan: Interesting question</li>
 <li id="l-158"><a href="#l-158">#</a> [17:09] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-159"><a href="#l-159">#</a> [17:09] -gitbot:#testing- [web-platform-tests] jgraham pushed 1 new commit to jgraham/html5lib_default_timeout: <a href="https://github.com/w3c/web-platform-tests/commit/ceb7717c3c3c1148089868ef5cf522e4f122c446">https://github.com/w3c/web-platform-tests/commit/ceb7717c3c3c1148089868ef5cf522e4f122c446</a></li>
 <li id="l-160"><a href="#l-160">#</a> [17:09] -gitbot:#testing- web-platform-tests/jgraham/html5lib_default_timeout ceb7717 James Graham: fixup! Remove the legacy timeouts from html5lib tests.</li>
 <li id="l-161"><a href="#l-161">#</a> [17:09] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-162" _a="jgraham" _r="zcorpan"><a href="#l-162">#</a> [17:09] &lt;jgraham&gt; zcorpan: Let's assume that it's good to be consistent (pushed a new commit to that review)</li>
 <li id="l-163" _a="zcorpan"><a href="#l-163">#</a> [17:10] &lt;zcorpan&gt; r+</li>
 <li id="l-164" _a="jgraham"><a href="#l-164">#</a> [17:12] &lt;jgraham&gt; Thanks!</li>
 <li id="l-165"><a href="#l-165">#</a> [17:12] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-166"><a href="#l-166">#</a> [17:12] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1199: Remove the legacy timeouts from html5lib tests. (master...jgraham/html5lib_default_timeout) <a href="https://github.com/w3c/web-platform-tests/pull/1199">https://github.com/w3c/web-platform-tests/pull/1199</a></li>
 <li id="l-167"><a href="#l-167">#</a> [17:12] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-168"><a href="#l-168">#</a> [17:12] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-169"><a href="#l-169">#</a> [17:12] -gitbot:#testing- [web-platform-tests] jgraham pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/7a53a49e4ad0f9d34680b2ba08af31cf62eaf103">https://github.com/w3c/web-platform-tests/commit/7a53a49e4ad0f9d34680b2ba08af31cf62eaf103</a></li>
 <li id="l-170"><a href="#l-170">#</a> [17:12] -gitbot:#testing- web-platform-tests/master 7a53a49 jgraham: Merge pull request #1199 from w3c/jgraham/html5lib_default_timeout...</li>
 <li id="l-171"><a href="#l-171">#</a> [17:12] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-172"><a href="#l-172">#</a> [17:24] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-173"><a href="#l-173">#</a> [17:24] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-174"><a href="#l-174">#</a> [17:31] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-175"><a href="#l-175">#</a> [17:45] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
 <li id="l-176"><a href="#l-176">#</a> [18:07] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-177"><a href="#l-177">#</a> [18:11] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-178"><a href="#l-178">#</a> [18:18] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-179"><a href="#l-179">#</a> [18:26] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-180"><a href="#l-180">#</a> [19:27] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-181"><a href="#l-181">#</a> [19:51] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-182"><a href="#l-182">#</a> [19:57] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-183"><a href="#l-183">#</a> [20:30] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-184"><a href="#l-184">#</a> [20:45] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-185"><a href="#l-185">#</a> [21:39] * Quits: ArtB (~abarsto@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-186"><a href="#l-186">#</a> [21:50] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-187" _a="plh"><a href="#l-187">#</a> [21:50] * plh is now known as plh-home</li>
 <li id="l-188"><a href="#l-188">#</a> [22:02] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-189" _a="plinss"><a href="#l-189">#</a> [22:15] * plinss is now known as plinss_away</li>
 <li id="l-190" _a="plinss_away"><a href="#l-190">#</a> [22:15] * plinss_away is now known as plinss</li>
 <li id="l-191"><a href="#l-191">#</a> [22:16] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-192" _a="abarsto"><a href="#l-192">#</a> [22:16] * abarsto is now known as ArtB</li>
 <li id="l-193"><a href="#l-193">#</a> [22:24] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-194"><a href="#l-194">#</a> [22:29] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-195"><a href="#l-195">#</a> [22:31] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-196"><a href="#l-196">#</a> [22:33] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-197"><a href="#l-197">#</a> [22:36] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-198"><a href="#l-198">#</a> [22:37] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-199"><a href="#l-199">#</a> [22:47] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-200"><a href="#l-200">#</a> [22:53] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-201"><a href="#l-201">#</a> [23:02] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-202"><a href="#l-202">#</a> [23:06] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-203"><a href="#l-203">#</a> [23:45] * Quits: plh-home (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-204"><a href="#l-204">#</a> [23:55] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-205"><a href="#l-205">#</a> Session Close: Wed Aug 20 00:00:00 2014</li>
</ol>
<p id="end">The end :)
